refactor: separate tests into different files
